requirements.txt
streamlit
seaborn
plotly
psycopg2-binary
python-dotenv
requests
beautifulsoup4
black
